Refactor the main file into distinct sources
