+++ README.md
@@ -1,1 +1,15 @@
+# Discord Meme Bot
+
 WIP but it is working!
+
+## Prerequisites
+
+- [Golang 1.20](https://tip.golang.org/doc/go1.20)
+
+## How to run
+
+```shell
+~ export DISCORD_TOKEN=<discord_token>
+~ go mod tidy
+~ go run main.go
+```
